sql_clinical/requirements.txt
requests
flask
flask_sqlalchemy
git+git://github.com/open-policy-agent/rego-python@master
